requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
setuptools==68.2.2
djoser==2.2.0
django-model-utils==4.5.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
flake8==5.0.4
drf-spectacular==0.24.1
gunicorn==20.1.0